Move inputs to inner folder
